+++ R/spatial_clustering_cv.R
@@ -9,17 +9,23 @@
 #'
 #' @details
 #' The variables in the `coords` argument are used for k-means clustering of
-#'  the data into disjointed sets, as outlined in Brenning (2012). These
+#'  the data into disjointed sets, as outlined in Brenning (2012), or for
+#'  hierarchical clustering of the data. These
 #'  clusters are used as the folds for cross-validation. Depending on how the
 #'  data are distributed spatially, there may not be an equal number of points
 #'  in each fold.
 #'
-#' @param data A data frame.
+#' @param data A data frame or an `sf` object (often from [sf::read_sf()]
+#' or [sf::st_as_sf()]), to split into folds.
 #' @param coords A vector of variable names, typically spatial coordinates,
 #'  to partition the data into disjointed sets via k-means clustering.
+#'  This argument is ignored (with a warning) if `data` is an `sf` object.
 #' @param v The number of partitions of the data set.
-#' @param ... Extra arguments passed on to [stats::kmeans()].
-#' @export
+#' @param cluster_function Which function to use for clustering. Must be one of either
+#' "kmeans" (to use [stats::kmeans()]) or "hclust" (to use [stats::hclust()]).
+#' @param ... Extra arguments passed on to [stats::kmeans()] or
+#' [stats::hclust()].
+#'
 #' @return A tibble with classes `spatial_cv`, `rset`, `tbl_df`, `tbl`, and
 #'  `data.frame`. The results include a column for the data split objects and
 #'  an identification variable `id`.
@@ -32,17 +38,43 @@
 #' pp. 5372-5375, doi: 10.1109/IGARSS.2012.6352393.
 #'
 #' @examples
-#' data(ames, package = "modeldata")
-#' spatial_clustering_cv(ames, coords = c(Latitude, Longitude), v = 5)
+#' data(Smithsonian, package = "modeldata")
+#' spatial_clustering_cv(Smithsonian, coords = c(latitude, longitude), v = 5)
+#'
+#' smithsonian_sf <- sf::st_as_sf(Smithsonian,
+#'                                coords = c("longitude", "latitude"),
+#'                                # Set CRS to WGS84
+#'                                crs = 4326)
+#'
+#' # When providing sf objects, coords are inferred automatically
+#' spatial_clustering_cv(smithsonian_sf, v = 5)
+#' # Can use hclust instead:
+#' spatial_clustering_cv(smithsonian_sf, v = 5, cluster_function = "hclust")
+#' @rdname spatial_clustering_cv
 #' @export
-spatial_clustering_cv <- function(data, coords, v = 10, ...) {
-  coords <- tidyselect::eval_select(rlang::enquo(coords), data = data)
+spatial_clustering_cv <- function(data, coords, v = 10, cluster_function = c("kmeans", "hclust"), ...) {
+  cluster_function <- rlang::arg_match(cluster_function)
 
-  if (is_empty(coords)) {
-    rlang::abort("`coords` are required and must be variables in `data`.")
+  if ("sf" %in% class(data)) {
+    if (!missing(coords)) {
+      rlang::warn("`coords` is ignored when providing `sf` objects to `data`.")
+    }
+    coords <- sf::st_centroid(sf::st_geometry(data))
+    dists <- as.dist(sf::st_distance(coords))
+  } else {
+    coords <- tidyselect::eval_select(rlang::enquo(coords), data = data)
+    if (is_empty(coords)) {
+      rlang::abort("`coords` are required and must be variables in `data`.")
+    }
+    coords <- data[coords]
+    dists <- dist(coords)
   }
 
-  split_objs <- spatial_clustering_splits(data = data, coords = coords, v = v, ...)
+  split_objs <- spatial_clustering_splits(data = data,
+                                          dists = dists,
+                                          v = v,
+                                          cluster_function = cluster_function,
+                                          ...)
 
   ## We remove the holdout indices since it will save space and we can
   ## derive them later when they are needed.
@@ -59,16 +91,31 @@ spatial_clustering_cv <- function(data, coords, v = 10, ...) {
     attrib = cv_att,
     subclass = c("spatial_clustering_cv", "rset")
   )
+
 }
 
-spatial_clustering_splits <- function(data, coords, v = 10, ...) {
+spatial_clustering_splits <- function(data, dists, v = 10, cluster_function = c("kmeans", "hclust"), ...) {
+
+  cluster_function <- rlang::arg_match(cluster_function)
+
   if (!is.numeric(v) || length(v) != 1) {
     rlang::abort("`v` must be a single integer.")
   }
 
   n <- nrow(data)
-  clusters <- kmeans(data[coords], centers = v, ...)
-  folds <- clusters$cluster
+
+  folds <- switch(
+    cluster_function,
+    "kmeans" = {
+      clusters <- kmeans(dists, centers = v, ...)
+      clusters$cluster
+    },
+    "hclust" = {
+      clusters <- hclust(dists, ...)
+      cutree(clusters, k = v)
+    }
+  )
+
   idx <- seq_len(n)
   indices <- split_unnamed(idx, folds)
   indices <- lapply(indices, default_complement, n = n)